refactorizando

+++ LIBRERIAS_Y_UTILIDADES/sources/lib/Collection/Collection.hpp
@@ -1,6 +1,8 @@
 #ifndef COLLECTION_HPP
 #define COLLECTION_HPP
 
+#include <cstddef>
+
 template <typename E>
 class Collection {
     
@@ -8,7 +10,7 @@ class Collection {
         // Returns true if this collection contains no elements
         virtual bool isEmpty() const = 0;
         // Returns the number of elements in this collection
-        virtual int size() const = 0;
+        virtual size_t size() const = 0;
         // Removes all elements from this collection
         virtual void clear() = 0;
 };
